I made this app for my software internship. This app has a connection with mssql.I used Rest API for listing items and upload item stocks.

+++ StokProgramiApp/Product.cs
@@ -12,7 +12,7 @@ namespace StokProgramiApp
         public string Name { get; set; }
         public int id_product_type { get; set; }
         public string serial_no { get; set; }
-        public double product_stock { get; set; }
+        public int product_stock { get; set; }
 
     }
 }
